Allow specifying absolute paths (with drive) on class attributes.
As long as the path is in fully-qualified DOS pathname format (with drive
letter prefix), the file path will be assumed as absolute and the file
specified in class' [FileNameI()] decorator will be created on the desti-
nation. Missing directories will be created recursively.

The path may be in the formats:
[FileName("C:\\Project\\folder1\\example.js")]
[FileName("c:/Project/folder1/example.js")]

Related GitHub issue: [bridgedotnet/Bridge#57]