Class' attribute 'Namespace' can now determines the file name
When using Namespace and NamespacePath as 'outputBy', having a class
with the [Namespace()] attribute  will determine the output filename
by its value. Previously it  was ignored and value from the assembly
was used.

This option does NOT take precedence over the [FileName()] attribute
if determined either on the class or on one of its superclasses!

[Fixes issue bridgedotnet/Bridge#104]