Addresses a translator issue with Mono and Project reading.
Microsoft.Builder.Evaluation.Project(string) overload has an issue in Mono
where it does not read the file at all.

The bug has been detailed in:
http://bugzila.xamarin.com/show_bug.cgi?id=38224

As an approach that works on both Windows and Linux, the correct overload
for the project is chosen just by changing the call from:
Project(fileName)

to:
Project(XmlReader.Create(fileName))

Which is anyway the chosen overload when just the first form is chosen.

The issue in Bridge is triggered by
Bridge.Translator.ProjectReader.GetSourceFiles()
Called when Bridge builder is about to read the list of source files inside
the .csproj project file.

Current Mono version always returns an empty list of files without the fix
this commit introduces.

This issue has been introduced with #531, released on Bridge 1.10.0, and
reported on #860, which may be released to 1.11.0 or 1.10.5.