Upgrades Bridge.NRefactory package references to 5.5.3.
Bridge.NRefactory has been updated to 5.5.3 so it uses Mono.Cecil
0.9.6.1. While the former was using an older Mono.Cecil conflicts on
assembly were happening on Bridge project. It could build on both
windows and linux but when used to compile a Bridge project on Linux it
failed due to the older version package's conflict and missing function.

This addresses issue #1030.