Fixes project OutputPath reading on Translator.
Translator.ProjectReader.cs now uses Microsoft.Build.Evaluation.Project
to read the OutputPath setting from the project file as well. It now
properly considers if the setting is limited by a conditional
(configuration, for example) block, getting the effective OutputPath
depending on the project's settings and chosen configuration.

This addresses an error reported on issue #1125 where the project file
had OutputPath defined on an non-conditional PropertyGroup.

It also refactors project reading mechanism so that the project is
opened just once during project read so that GetSourceFiles() and
GetOutputPath() uses the same loaded instance of the project file.

Additionally, the change in parameters to GetSourceFiles() required
renaming both overloads of the function to GetProjectSourceFiles() and
GetFSSourceFiles().